Add files via upload
The project we are going to do is to create a calculator using HTML, CSS, and JavaScript.The prerequisites of this project are the basic knowledge of HTML CSS and javaScript. Here are the project details:

Project Explanation:-
As we know that a calculator is a machine which allows us to do math operations more easily. For example, most calculators will add, subtract, multiply, and divide. Some also do square roots, and more complex calculators can help with calculus and draw function graphs. Calculators are found everywhere. You can even build a more advanced level calculator, but we will keep it simple.
Start by creating a new project folder in VS Code, and inside that project folder, create three empty new files of HTML, CSS and javaScript. The Html file is the entry point for our website and contains the HTML code. And the CSS file designs of the content of the HTML. Whereas the javascript add functionality in the calculator.